requirements.txt
torch
torchvision
segmentation-models-pytorch
opencv-python
numpy
pyyaml
wandb
tqdm
hydra-core
scipy